Weather_ETL_Pipeline/requirements.txt
requests
pandas
mysql-connector-python
python-dotenv
